Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > This isn't actually ready for review, just wanted some eyes and suggestions on how it should look.
> > > 
> > > https://dev-simonhatch-c1a8f15f-dot-pinpoint-dot-chromeperf.appspot.com/job/138e4659ec0000

When I open this on a laptop and click SEE ALL in both columns at the bottom, the text overlaps and is difficult to read. Maybe add this CSS?
#exception {
  overflow: auto;
}

Add the 'raised' attribute to the SEE ALL button so it looks more like a button than the surrounding text and maybe give it a background-color to match the nav bar.

The expand-more icon would be more appropriate than the add icon. Change it to the expand-less icon when expanded. Move the icon to before the text.

"Show" is a more common UI verb than "see", but if you can use a more specific noun than "all" like "exception", then you can just use that noun without "show".
https://ell.stackexchange.com/questions/41671/show-details-vs-see-details

Maybe draw a border around exception-details when expanded to show the user what the expand-less button would hide? Maybe use a tab instead of a button?

Who knew buttons could be this complicated? :-)

> > > 
> > > Basically surfacing the exception as a human readable version, and then the full stack trace + any associated other data (such as telemetry stack trace) in the "see more" section. Could use some help making it look not terrible.
> > 
> > I think the CL is missing the exception-details.html file, but the UI looks fine to me.
> 
> Some feedback after having a very quick re-look:
> 
> - We should give an option to hide the details again.
> - We can also differentiate the colour/presentation of the exception, maybe with a different font (monospace maybe) and slightly lighter colour (maybe dark/medium grey).

+1 to Dean's suggestions. Maybe grey, maybe red.

Patch-set: 1
Reviewer: Gerrit User 1115874 <1115874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>