PinpointScheduler: Integrate a FIFO Scheduler
In this change we implement a simple FIFO scheduler for Pinpoint, which
intercepts all new jobs created and attempts to isolate those for the
same "configuration" (an identifier for a group of machines in the
Swarming pool), and runs each one in insertion order.

We leave a bunch of TODO's which indicate how we'll implement a more
full-featured scheduler as a service, independent of the Pinpoint
service, but have a `scheduler` model which we'll operate through.

In the scheduler, we support the following APIs with documented
semantics, but at a high-level:

  * `scheduler.Schedule` takes a `Job` instance and identifies which
    FIFO queue it should be appended to.
  * `scheduler.PickJob` provides the first queued or running job for a
    given configuration.
  * `scheduler.Cancel` marks a Job for cancellation.
  * `scheduler.Complete` marks a Job as completed.
  * `scheduler.Remove` forcibly removes a job from a queue by ID.

We currently also leave behind markers in unittests for supporting
cancellation and how to handle jobs stuck in the "running" state which
we don't have implementations associated with in the user interface yet.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:977488
Change-Id: Ife9192fabc492fb7524a1a4abd99aea2bde49246