Generate reports in the spa frontend.
There are currently several issues with loading reports.
The main problem is that reports are generated in the server
under a strict 60s timeout, which is frequently hit.
This CL removes /api/report/generate in favor of generating reports
in the client. Depending on the report template, this can cause
several hundred more requests to /api/timeseries2. However, thanks to
the service worker and moving logic to the client, this change should
save a significant amount of appengine CPU overall.

This change also allows removing ReportCacheRequest and ReportRequest,
reducing the amount of code under dashboard/ by over 1kLoC.

ReportMerger maintains the feature of ignoring test cases that were
added to or removed from the benchmark during the revision range
configured by the user via the ReportControls.

This CL also fixes td.rowspan in report-table.

This CL also makes TimeseriesCacheRequest remember which timeseries
return 404 and avoid re-fetching them for 2.8 days so that it will
check to see if they have been added on the third day.

This CL also makes TimeseriesCacheRequest retry 500s up to 3 times.
Relatedly, servicifying /api/* should allow tweaking the YAML settings
to handle more api requests in parallel:
https://chromium-review.googlesource.com/c/catapult/+/1660976

This CL also removes the tooltip from ReportTable. Instead of hovering
over a row and scanning for a test case, users can click the blue label
to open a chart section and open the Case menu: if a test case was
merged into the report, then it would be checked in the Case menu.
This CL also changes TimeseriesRequest to handle multiple statistics.

Test coverage is maintained at 81% Stmt, 72% Branch, 84% Func, 84% Line.

Followups could refactor a ChartFetcher out of ChartTimeseries,
refactor ChartFetcher to share code with DetailsFetcher and ReportFetcher,
and refactor details-fetcher's RangeFinder to share code with
report-fetcher's RangeFinder.

Demo: https://v2spa-dot-chromeperf.appspot.com/

Bug: chromium:971655

Change-Id: Idfb86fe48a6f72756ad2bdf6496da21bbcaad6bf