Update patch set 45
Patch Set 45:

Had a chat offline with Ben. This may not exactly fit our purposes but: (1) it's strictly speaking an improvement over the current situation, and (2) exactly what "our purposes" are is still somewhat in flux at the moment.

So, do go ahead with the CL. I'll chat with others to try to have some more concrete thoughts on how a "system health rotation" would look like.

Patch-set: 45