Store artifacts of each story run in a separate directory
This CL changes the interface of functions dealing with artifacts a bit.
Now 'name' of an artifact represents the actual file name inside the
artifacts folder, which is different for every story run.

Also there's now an 'Artifact' object that stores local and remote path
to the artifact along with its content type.

Also a CaptureArtifact method is added for users who want to create
a file on their own.

For a discussion of artifact representation, see [1].

[1] https://docs.google.com/document/d/1TOalrWznY1II3L_6bC7lzebbOxYDBzkf3LCXr3XJEBs/edit?usp=sharing

Bug: chromium:972025
Change-Id: I1ba26713fe6774af36f4e7c857ee101d262d185a