Inline chart options.
Chart options allow users to link charts together, zero y-axis, fix
x-axis, and select a normalization mode.
Currently, chart options are in an abspos dialog displayed by clicking
the gear icon to the left of the minimap.
There have been a few questions indicating that this dialog is not
discoverable enough.
This CL moves these options into an inline box with a chevron pointing
to the gear icon, so it can be displayed by default and easily hidden.

Change-Id: Ib5b56021c8ccc247de4289673fdc14f6060ff881