Pinpoint - Update the error UI
This separates the messaging and the full stack + associated other
details (such as telemetry exception) in the UI. The human readable
messaging is displayed by default, with an option to expand and see the
full dev stack trace and any extra info plumbed up.

In the future, we may include more information like markdown with
remediation steps, data for how to file a bug, etc.

Bug: chromium:971370
Change-Id: Ib24b5cc14354afce17011ffda9a7f83b535152b2