Update patch set 1
Patch Set 1:

We discussed the issue offline and agreed that losing exceptions is a worse problem than logging them twice.

Juan, should we land this as it is as a quick fix, before the broader refactoring is ready?

Patch-set: 1
Reviewer: Gerrit User 1328067 <1328067@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>