[Telemetry] Integrate new test expectations format into telemetry benchmark runner
We are attempting to standardize the test expectations format. Telemetry benchmark
runner currently uses a format that is different than the standard format. This CL
will make it support the new format as well as  the old format. After we make
modifications to the expectations file for chrome performance benchmarks so that
it uses the new format, we will remove support for the old format. This CL also
includes a bug fix for Typ. When no reason is given for a test expectation, then a
None type value should not be included in the reasons set for the expectations of a test.

Bug: chromium:973936
Change-Id: Ic64239cd8a275c42f51742b6b45dc15e100865c1