Display icons in charts for osVersions changes.
Seeing when certain diagnostics change can help users understand
changes in the data. Currently, charts only display icons for alerts.
At one point, charts also displayed icons for all changes in the
diagnostics whitelisted here:
https://cs.chromium.org/chromium/src/third_party/catapult/dashboard/dashboard/add_histograms.py?l=33
That was too noisy, so it was removed.

This CL adds a new whitelist of diagnostic names to display icons
in charts as an experiment. If this is too noisy, it may be removed.
If not, other diagnostic names may be added.

Bug: chromium:980256
Change-Id: I97b6aae07a71b3a8aca65b2e4d9d26bb1ac5233c