[Telemetry] Clarify difference between story_set and list of stories.
story_filter.py has always taken in an iterable of stories and output
a list of stories, but based on the documentation, readers would
suspect that it takes a story_set and returns a story_set.

Similarly for the story_runner code, this difference is not clear,
and that leads to a lot of confusion.

Bug: chromium:965158

Change-Id: Idbba727df0bdd5c0b9bfff8737077b9c6f8bfab2