Pinpoint - Check new completed field for results2
Results2 files were failing to get generated for try jobs because we
were checking if the job had a task, which previously marked whether the
job was complete or not. There's a proper "completed" field since the
introduction of the scheduler, since a job can also not be started
yet.

Bug: chromium:980887
Change-Id: I03eee78d1e6c9a07d3b30ba239241ea3bbc9ec38