PinpointScheduler: Expose FIFO Queue Stats
This change implements the queue stats computation which allows us to
expose this information in a user-visible queue stats page. This change
also includes the rudimentary presentation page for showing stats for a
specific queue.

We also add a link to the queue stats page from the job table, so that
users can have a more direct access to seeing statistics about a queue.

A future change will allow us to get information about all non-empty
queues for a consolidated view for the statistics in the /stats page.

Another potential future change should allow us to make the stats for a
given queue be loaded as an overlay, without having to bring users to
the stats page for that queue.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:977488
Change-Id: Ie505d14308e0170f9448c4ad0f049a7462b5ddb6