Update patch set 1
Patch Set 1:

(2 comments)

Are you planning to take this further and rename page to story throughout telemetry?

If I understand correctly, a Story was thought to be an abstraction for testing applications other that Chrome, but no other subclasses of it were implemented apart from Page. So all Stories that we deal with are actually Pages, and technically both "story" and "page" are correct names for them.

I think it might be possible to rename everything automatically after we move telemetry to chromium, but before that we're stuck with this ambiguity.

Patch-set: 1
Reviewer: Gerrit User 1328067 <1328067@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>