Update patch set 1
Patch Set 1:

(2 comments)

> Patch Set 1:
> 
> (2 comments)
> 
> Are you planning to take this further and rename page to story throughout telemetry?

There was a plan to eventually do this (although I can't find the bug now). I don't plan to finish the renaming through all of telemetry any time soon but, specially for new code, I think it's good to switch to the new terminology whenever we can. That's why I wanted to do the renaming here before this other upcoming CL:
https://chromium-review.googlesource.com/c/catapult/+/1688799

> If I understand correctly, a Story was thought to be an abstraction for testing applications other that Chrome, but no other subclasses of it were implemented apart from Page. So all Stories that we deal with are actually Pages, and technically both "story" and "page" are correct names for them.

Indeed story was introduced as a generalization of page which, at least conceptually, is not necessarily tied to a single "page" or URL. And it's true that most stories are pages, but not all of them are. Notably the startup_mobile [1] benchmark which does more interesting non-page things subclasses directly from Story, so I think this is a good abstraction to have.

[1]: https://cs.chromium.org/chromium/src/tools/perf/benchmarks/startup_mobile.py

Patch-set: 1
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0