Don't use UE bucketing in runtimeStatsTotalMetric
We used to bucket runtimeStatsMetric based on RAIL stages. The idea
behind this was to use the load bucket as a way to measure the loading
time. Though this was always noisy and was never used a lot. Also,
importing a trace and computing the metrics is expensive and is causing
timeouts on some bots. Avoiding UE bucketing would avoid the need of
some trace events and also reduce the cost of computing metrics.

This cl also adds a V8-Only-Foreground metric to see the amount of time
spend on mainthread in V8.

Bug: chromium:958441
Change-Id: I4784f4ac4f15af418d27181925b2bf87fbf11f37