Devil: support Q in SetWebViewFallbackLogic
SetWebViewFallbackLogic is supposed to NOOP for API levels which don't
support the command. The command was ripped out in Q, so this must
early-out for >= P devices. This adds a test case to check this API
level works correctly.

I realized the API doc was incorrect (it talked about _switching_
implementation, rather than fallback logic), so I also updated the API
doc with more context on the complicated "fallback logic."

Bug: chromium:948905
Test: ./devil/bin/run_py_tests
Change-Id: I08cf9abf45178b43b6cb5d8a73f4f11cad799cdb