[Telemetry] Fix benchmark interruption logic.
The primary motivation for this changelist is to
unify the story filtering logic into one place.
Once this changelist is submitted, it will be
easier to put all the story selection logic
into a single place per
https://bugs.chromium.org/p/chromium/issues/detail?id=982027#c2
So mostly this is a cleanup, but it does involve behavior
changes, which I would consider to be bug fixes.

Behavior changes:

1. If you set a story to be skipped in expectations.config
file, then it will be marked as expectedly skipped whether
the benchmark gets interrupted (unhandleable error)
or not. (i.e. fixes crbug.com/980781)

2. If you set max failures, then this changelist will mark
any story after the max failures as unexpectedly skipped
(rather than pretending that those stories did not exist).

bug: chromium:980781
Change-Id: I894119585008e2303b944c53a3c6d2b451cb61b1