Pinpoint: Stop relying on device affinity
This change removes the requirement from Pinpoint jobs to run on the
same devices that previous tests have executed on. We're doing this to
reduce the occurrence of devices going offline causing attempts to be
re-done, and from swarming tasks waiting for a specific device to be
available.

We determined that the trade-off of potential variance due to device
differences to be sufficiently covered by Pinpoints attempt scaling in
response to variance in data. We're trading off the perceived stability
introduced by device affinity for better resource utilisation and
resilience to device unavailability.

R=simonhatch@chromium.org

Bug: chromium:948251
Change-Id: I105d89c4f30f1be5f596e4ad8a34d7bbb0845135