cp-flex simplifies flexbox in spa.
V2spa uses display:flex a lot. Most uses are for divs or spans, which are
semantically empty.
This change defines an element for these uses. This reduces CSS and
allows the HTML to be clearer about its layout.
A few attributes are defined for common settings such as
flex-direction:column and flex-grow:1.

No functional changes.

Change-Id: I8f597faf0a59397b412f6f256ce6a37d76e267fb