[Telemetry] Promote story run independence on all platforms.
With the removal of memory.top_10_mobile; we no longer need to support
benchmarks intentionally breaking story independence which was still
allowed on Android.

This allows to simplify the control flow in WillRun/DidRunStory. Also
deprecate ShouldStopBrowserAfterStoryRun and introduce a new simpler
ShouldReuseBrowserForAllStoryRuns method for the few remaining
legitimate use cases.

Bug: chromium:983172
Change-Id: I2eb0a4de6ad96272248d09150bf31dfb87884521