Improve system tracing failure error message
If system tracing is enabled then new and exciting failure modes are
enabled. One of the most obscure is that if DisablePagingExecutive is
not set then tracing may be unable to determine whether it is running
on a 64-bit system. The fix is quite non-obvious.

In addition to suggesting that the user set DisablePagingExecutive I
also suggest that they disable System tracing because that is simpler
and system tracing is often not actually needed.

Bug: chromium:604301
Change-Id: I09a73fdb4fb01e6835aa1e8999936cc6df88221e