[devil] Fix parsing of surface flinger data
Skip over and warn about lines with non-numeric data. On nexus5x the
command to gather surface flinger data is sometimes including in the
output:

  SurfaceFlinger appears to be unresponsive (Connection timed out
  [-110]), dumping anyways (no locks held)

We want to skip over this and attempt to parse the rest of the output.

This CL also separates out the parsing code to its own function so
it's easier to test.

Bug: chromium:979834
Change-Id: Ic2454127e527a08813a68b3a81f1aaf8cf5940f2