[Telemetry] Split benchmark_runner.main into ParseArgs and RunCommand
This is a first step towards allowing Telemetry to be combined with
an external results_processor module; openining up components which
the entry point script (run_benchmark) can control and combine.

Also take the chance to introduce a new top level argument parser,
implemented using argparse, rather than the deprecated optparse. This
is needed to avoid spreading the dependence of this deprecated module
on new code for the results_processor.

Bug: chromium:928275
Change-Id: If62ffb29795f060d2ef3e9c2524cd5c03f7258bd