Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> https://dev-simonhatch-3f9e5872-dot-pinpoint-dot-chromeperf.appspot.com/
> 
> Hey, not ready for a real review, just take a poke at the UI and let me know what you think. Took a stab at replacing the text fields using dashboard's autocomplete box widgets, so we have a dropdown of suggestions on existing benchmarks/stories/storyTags instead of the text fields.

This looks cool!

I think this is the right direction.

There's a weird issue with attempting to grab the scrollbar (clicking it closes the drop-down for me) but otherwise this looks nice.

I'm a little uneasy about referring directly to the dashboard's elements, but I suspect it's fine for now (for simplicity of deployment).

Does it make sense to also allow folks to edit the attempt to re-run with the same UI, instead of assuming a re-run will always have the same arguments?

Patch-set: 4
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>