Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > https://dev-simonhatch-3f9e5872-dot-pinpoint-dot-chromeperf.appspot.com/
> > > 
> > > Hey, not ready for a real review, just take a poke at the UI and let me know what you think. Took a stab at replacing the text fields using dashboard's autocomplete box widgets, so we have a dropdown of suggestions on existing benchmarks/stories/storyTags instead of the text fields.
> > 
> > This looks cool!
> > 
> > I think this is the right direction.
> > 
> > There's a weird issue with attempting to grab the scrollbar (clicking it closes the drop-down for me) but otherwise this looks nice.
> > 
> 
> Wonder if Ben has some suggestions here. Feels like if we keep the menu around it becomes hard to dismiss.
> 

+1 -- although if I think about it, I suspect this is fine for now. Optionally, hiding the scrollbar might be an option? That would reduce the impulse to click on it if it doesn't show up.

> > I'm a little uneasy about referring directly to the dashboard's elements, but I suspect it's fine for now (for simplicity of deployment).
> > 
> 
> Yeah just for prototyping, like you say it's just easier. I'm hoping eventually these codebases are brought together anyway.
> 

That's a thought -- the underlying data provider for this is the API, right? How hard would it be to consolidate the elements into a separate service (or set of static files) that we could be using across the dashboard and Pinpoint?

I understand that we're already currently shipping these already, but having a central place for the components just sounds like a good practice anyway. The alternative is if we start vulcanising these components instead so that we can reduce the actual look-ups and package down "optimised" versions of these components that we use.

I looked at this briefly and I think if we introduced makefiles that allow us to run the vulcanize tool (or maybe have it in the deploy.py script as well) to generate the vulcanised components for Pinpoint, then we may be able to get away with not having to share the web component definitions at all.

That sounds like a larger project than we can take on depending on how you feel about addressing this deployment issue as well, but it's certainly an option we can explore to avoid having to do these gymnastics just to get code sharing going on.

> > Does it make sense to also allow folks to edit the attempt to re-run with the same UI, instead of assuming a re-run will always have the same arguments?
> 
> Definitely think we should allow editing of the params, but this isn't the rerun UI (it's the new job UI) so I think we should just make that a follow up.

Right -- I agree with the follow-up suggestion. But it does beg the question why those are different UIs (when in my mind it feels like one is a subset of the other, e.g. the rerun is like this new job UI but with fields pre-filled).

Patch-set: 4
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>