[Telemetry] Implement new top level parser
This new top level parser, implemented using argparse, for the moment
just handles choosing a command (either list or run), and hands of
parsing of the rest of the args to the specific parser of the chosen
command.

The old "Help" command is also removed, relying instead on the default
behaviour of the --help flag.

Tests are added to make sure that various common command line patterns
are still parsed and interpreted correctly.

Bug: chromium:928275
Change-Id: I328892f7a4e29e057d7712bf7bd7d86a11ba5067