[Typ] Augment test expectations format by adding a new 'conflicts_allowed' descriptor
If the descriptor value is set to true then the expectations parser will check for conflicts
between test expectations. For example the following expectations would conflict.

[ win intel ]  FooTest.TestBar [ RetryOnFailure ]
[ win ] FooTest.* [ Failure]'

This would be problematic because a regression would leak. It would waste CPU cycles to find.
And someone would have to manually find the conflicting expectation and then remove it.

Bug: chromium:973936
Bug: chromium:976505
Change-Id: I79dfb52feb6bf9d78e88bd7fd616d35abb2e8e86