Update patch set 26
Patch Set 26:

(4 comments)

> Patch Set 26:
> 
> > Patch Set 26:
> > 
> > Biggest comment right now is the concern over complexity here. Don't feel like we need to have the the super general, can attach notes across bots/benchmarks/whatever unless it's explicitly requested, which will simplify this cl.
> 
> Tdresser notes on the bug that he expects to use notes for metric changes, which would apply across all bots and cases, and possibly suites.
> 

Fair enough, I wanted to be sure it was addressing a real need.

> > 
> > Also, have you done any demo showings of the new notes UI?
> 
> Tdresser gave UI LGTM on the bug.
> 

Works for me.

> I could solicit feedback from sadrul and other users if necessary?
> 

If you want to, go for it. I wanted to be sure a few people have taken it for a spin first. At the time of the question, there wasn't much feedback in the bug.

> Offline, you mentioned a concern about migrate_test_paths not updating notes. Should we discuss possible resolutions for that problem? Any other concerns?

Yes, definitely needs to be migratable. Separately I think Juan is proposing some big changes to story naming in telemetry, which may see a lot of renames go by when that happens and these notes will almost certainly get caught in that.

Patch-set: 26
Reviewer: Gerrit User 1116018 <1116018@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>