Add ReportFetcher to spa.
Currently, reports are generated on the server. However, large reports
often hit appengine's 60s timeout.
ReportFetcher will be used to generate reports on the client.
It takes a template report and a list of revisions, fetches timeseries
data, collates it, and yields it to the caller.
This new class is similar to DetailsFetcher and ChartFetcher.

Large reports often fetch data for hundreds of timeseries. In order to
maintain low latency, this fetcher actively manages revision ranges to
ensure that it fetches as little data as possible so that each request
to the backend can be as fast as possible.

This CL is split from
https://chromium-review.googlesource.com/c/catapult/+/1649350

Bug: chromium:971655
Change-Id: I0d1e3e3cda5a98ea42777057b53e9ceca346f344