Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (3 comments)
> 
> I'm not sure these stream hacks are correct. shouldn't we be processing bytes/str based on the open mode? if a steam is opened as binary, we shouldn't be allowing write calls with strings.

Yeah, that's what I meant about needing to be careful. The interface provided through the host isn't a general-purpose mechanism that lets you control the file modes for stdout/stderr, so I think we can (and should) assume that we have strings. If there's a way to assert on this, that'd probably be a good idea.

Patch-set: 3
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>