Pinpoint: Only refresh running jobs
This change ensures that we only re-run jobs that are already running,
not those that have been queued for longer than the threshold.

Now that we have a scheduler, we can get to a state where a job that's
had task names assigned to them could be considered frozen even though
they're just "queued". We change the definition for the running
property to take into account whether a job has been scheduled in the
past.

Checking whether a job is not 'completed' is not a sufficient condition
thus we're changing the query to only look for and process the 'running'
jobs.

R=abennetts@chromium.org, simonhatch@chromium.org

Bug: chromium:984850
Change-Id: I788cacf0265374418ad778e9dfe4832098de247a