ChromePerf: Fix dispatch rules
It turns out the previous version of this file defined too broad a set
of routes, and we've had to constrain the routes to differentiate
amongst the services which also overlap some of the /api/* routes. This
caused a minor service disruption for both Pinpoint and the Sheriff
Configuration services.

R=abennetts@google.com, simonhatch@chromium.org

Change-Id: Ibdba7c47eefa8cc8c04d068c79ff674a06b88cbe