Update patch set 53
Patch Set 53:

A general comment, not exactly sure if related to this CL. If looking at multiple reports, while data is loading the reports and metrics jump up/down all over the place, aesthetically it looks very bad.

Also, not sure if there is a way to tell when loading of the reports is "done", and/or whether loading of some metrics failed entirely.

Patch-set: 53