[Telemetry] Move suppress_gtest_report out of result options
The 'gtest' is not really an output format (in the same sense that csv,
html, histograms) are. So this CL removes it from result_options.

For the same reason suppress_gtest_report is moved now into story runner
options. Now the CreateResults function, instead of directly accessing
this flag, gets an explicit "progress_report" flag which defaults to
False. This also makes it easier to uniformly disable progress reporting
during tests.

Bug: chromium:981349
Change-Id: I6aa6761fa2920133ffc07e3fb4192f7d068cb5fa