[Telemetry] Fix and re-enables android repeatable scroll tests
Adds repeatable scroll tests back on android. These were most
likely failing due to compatibility issues with blank.html not
working on mobile. This update will allow it scroll/be resized
correctly. The check also skips the test if the window could not
be resized - if there is no resize, then the scrollbar cannot
be expected to scroll past the end.

Bug: chromium:972169
Change-Id: I5329500f2c7b320d39f499191bfc62ee13a64ea4