[Telemetry] Add GetRecentMinidumpPathsWithTimeout
Adds GetRecentMinidumPathsWithTimeout to Telemetry's Browser and
BrowserBackend classes. This blocks until either the requested number of
recent dumps (configurable as how old the dump can be relative to the
current time) are found or the timeout is reached.

This is to work around the race condition where Telemetry looks for a
minidump after a crash but before the dump has actually been written to
disk.

Bug: chromium:634156
Change-Id: I296662846bf734de55f525c3d8c6a13650107198