[Telemetry] Add GetRecentMinidumpPathWithTimeout
Adds GetRecentMinidumPathWithTimeout to Telemetry's App and
BrowserBackend classes. This blocks until either a recent dump
(configurable by specifying the oldest allowable timestamp) is
found or the timeout is reached.

This is to work around the race condition where Telemetry looks for a
minidump after a crash but before the dump has actually been written to
disk.

Bug: chromium:634156
Change-Id: I296662846bf734de55f525c3d8c6a13650107198