Pinpoint - Simplified UI
This adds autocomplete boxes for a lot of the fields on the new job
UI. We use the dashboard's api to xhr out to get a list of benchmarks,
and their stories/tags. Story/StoryTags aren't
available until after a benchmark is chosen, since we need to xhr
to the backend to get the more info on the benchmark.

Bug: chromium:982391
Change-Id: I80b425997fc3963d54d2ce4a2e24e19c1c29c7a1