Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I wonder if the concept of "tags that conflict" is clearer than "tags that overlap"; similarly, I'm also not sure how to interpret your example in the description. 
> 
> I.e., "win" and "win7" overlap. They should conflict, right? But, "win" != "win7", so the actual implementation would say that they *don't* overlap?

win and win7 shouldn't conflict because both win7 and win tags apply to a system with the windows 7 OS. A system with the windows 7 OS is also a windows system. Tags win8 and win7 would conflict because windows 7 and 8 are different operating systems. Its up to the developer to pass a function that equates the win7 and win tags. The lamba passed to parse_tagged_list is used to find tags are from the same tag declaration list and are not equal to each other by using the opposite of the return value of the lambda.

Patch-set: 2
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>