Update patch set 2
Patch Set 2:

The philosophy is that the win tag should apply to all windows systems which can have windows 7, 8, xp and all other windows operating systems. Its up to the test runner to include the generic OS name tag which is 'win' and specific OS version name tag which could be 'win7' in the  list of tags that describe the system which passed to Typ. This is so that any expectation with a generic OS tag like 'win' and specific version tag like 'win7' would be apply for the same test run.

Patch-set: 2
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>