[Telemetry] Require output_dir to CreateResults
Make it mandatory to supply an output_dir when calling CreateResults.

This directory is used to place any created artifacts and, soon, the
intermediate test results produced by Telemetry.

Although one could also support a mode with "no output", it's more
desirable to keep a single code path that all scripts and tests will
use.

Note: It is still possible to create the underlying PageTestResults
object with output_dir=None, and some tests do. That should be
addressed in follow up CLs.

Bug: chromium:979194
Change-Id: I249ad5e895ff36ae52932ca9688083fdfa18bbab