PinpointScheduler: Allow cancellation in job page
This adds the controls required for supporting cancellation in the
Pinpoint job page. We introduce two new badges (cancelled and queued) to
better surface the state of a job when looking at the details. Beside
some of the badges (running and queued) we place the 'X' allowing for
bringing up the cancellation dialog.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:977488
Change-Id: I24ce005ac86e07cb17ad662ccccf820a57df4fb9