[Telemetry] CreateResults in only one place.
Behavior changes include that we will create blank results output
files for cases where the benchmark was disabled on the platform.
I think this is desirable since it means that users can always
expect a well-formatted results file (even if we can't find
a possible browser to run!). And that means that in the future
if we choose to extend Telemetry to provide a "failure reason"
in structured output (for example Pinpoint has been asking for this)
then we could handle outputting that from page_test_results.py
for a greater set of cases.

Refactor means that _ShouldRunBenchmark does not have side
effects. I think given the function's name that side effects
are not a good idea. Also,
CreateResults is only called in one place, which is simpler.


Bug: chromium:985103
Change-Id: I5f462a13904a298dd5939f0b090097bb575f3bd4