[Telemetry] Merge command line help from Telemetry and external parsers
When supplying an external results_arg_parser, Telemetry will now
combine both the help strings of its own option parsers, with the
ones defined in the external one.

To achieve this, processing of the --help flag is transfered away from
the legacy parsers and into the new top level parser.

Bug: chromium:981349
Change-Id: I4811cfcb15f6da6a63eaed71d66fa4e85fa6df50