[Telemetry] Merge command line help from Telemetry and external parsers
When supplying an external results_arg_parser, Telemetry will now
combine both the help strings of its own option parsers, with the
ones defined in the external one.

To achieve this, processing of the --help flag is transfered away from
the legacy parsers and into the new top level parser.

For a sample of how this would look like when a suitable external
results processor is used see:
https://gist.github.com/perezju/2a023ab7e08afb78f21bfa7c94a54090

Bug: chromium:981349
Change-Id: I4811cfcb15f6da6a63eaed71d66fa4e85fa6df50