[Telemetry] Add generic typ expectation tags for desktop and mobile platforms
These tags can be used to shorten the amount of expectations. Instead of
writing an expectation for each desktop or mobile platform, only one expectation
can be written using a generic tag. For instance the desktop tag can be used to
cover linux, mac, windows and chrome os tags. This CL also adds the android-webview
and android-not-webview tags which cover all webview and non webview android
configurations.

Bug: chromium:973936
Change-Id: Id7258698154d06b6abb1a6e0ebd3d9a22fec50db