Update patch set 1
Patch Set 1:

hey Juan, I am sorry for not updating the docs on a more timely manner, but I think we won't need this after all.

In https://docs.google.com/document/d/1nzqrwfTJtzU71no_IiOh80HYBIpr1OPoJJonQIJPWtI/edit I am trying to define how *full* test paths would look in Chrome, and of the pieces of proposal is to use gn/{target}/{test_name} as test path. I've learned/was reminded that GN targets have slashes and colons as hierarchy separators (e.g. //chrome/testing:browser_tests), and it would be weird for developers to use slash instead of colon/. So requiring slashes everywhere is not very practical. Also it requires more complexity to distinguish full target and full test name, e.g. gn/chrome/testing/+/browser_test/+/Class/Method. This is getting complicated.

I had to step back. This lead to an idea that, actually, for the problem we are trying to solve (more compact representation of repeated prefixes in UI), a formal hierarchy is only one of the solution. Another is, simply, recognize that 2 or more test paths have a common prefix, not necessarily ending with a particular symbol. For example,

  aaaa.bbb/ccc:ddd
  aaaa.bbb/ccc:fff

a simple UI algorithm can treat "ddd" and "fff" as children of "aaaa.bbb/ccc:" and display them as such. The only piece of complexity here is to choose a possible separators, e.g. [^:alphanum:].

So this is what I am going to propose once I am done with some things I'm currently busy with. This means slashes are not special, and its ok for telemetry to use them in test paths, and you don't have to escape them.

Patch-set: 1
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>